Add files via upload
The zipped files for breast cancer and brain tumor contain the final results of the proposed algorithms, which use the U-Net model combined with IQR. These results clearly demonstrate the model's ability to detect cancer and tumors in affected patients. The results section includes the IOU curve, accuracy curve, loss curve, and Dice coefficient curve for detailed analysis. Additionally, a .txt file is included with the evaluation metrics of the proposed model. The actual output masks generated by the model are also provided. Please note that the source code has not been shared, as it is confidential and currently under consideration for publication in a Springer Nature journal.